python/requirements.txt
bcrypt
fastapi
pymongo
python-dotenv
uvicorn